requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-community>=0.0.28
langchain-openai==0.0.8
chromadb==0.4.24
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
pypdf==4.1.0
tiktoken==0.6.0 